requirements.txt
streamlit
qdrant-client
sentence-transformers
groq
python-dotenv
Pillow
pandas
torch